requirements.txt
langchain==0.2.1
streamlit==1.35.0
uvicorn==0.30.1
python-dotenv=1.0.1
google-generativeai
PyPDF2==3.0.1
faiss-cpu==1.8.0
langchain_google_genai
